refactor: Player 삭제
미션 본연의 목적에 집중하기 위해,
게임에는 필요하지만 연관관계가 있는 Player 삭제
