+++ src/main/java/janggi/domain/game/Game.java
@@ -2,26 +2,23 @@ package janggi.domain.game;
 
 import janggi.domain.board.Board;
 import janggi.domain.board.setup.BoardSetUp;
-import janggi.domain.player.Player;
-import janggi.domain.player.PlayerSetUp;
 import janggi.domain.side.Side;
 
 public class Game {
-    private final PlayerSetUp choPlayerSetUp;
-    private final PlayerSetUp hanPlayerSetUp;
+    private final GameSetUp choGameSetUp;
+    private final GameSetUp hanGameSetUp;
     private final Board board;
 
-    private Game(PlayerSetUp choPlayerSetUp, PlayerSetUp hanPlayerSetUp, Board board) {
-        this.choPlayerSetUp = choPlayerSetUp;
-        this.hanPlayerSetUp = hanPlayerSetUp;
+    private Game(GameSetUp choGameSetUp, GameSetUp hanGameSetUp, Board board) {
+        this.choGameSetUp = choGameSetUp;
+        this.hanGameSetUp = hanGameSetUp;
         this.board = board;
     }
 
-    public static Game createGame(Player choPlayer, Player hanPlayer,
-                                  BoardSetUp choBoardSetUp, BoardSetUp hanBoardSetUp) {
+    public static Game createGame(BoardSetUp choBoardSetUp, BoardSetUp hanBoardSetUp) {
         return new Game(
-                new PlayerSetUp(choPlayer, Side.CHO, choBoardSetUp),
-                new PlayerSetUp(hanPlayer, Side.HAN, hanBoardSetUp),
+                new GameSetUp(Side.CHO, choBoardSetUp),
+                new GameSetUp(Side.HAN, hanBoardSetUp),
                 Board.setUp(choBoardSetUp, hanBoardSetUp));
     }
 }
