--- src/main/java/janggi/domain/player/Player.java
@@ -1,9 +1,0 @@
-package janggi.domain.player;
-
-public class Player {
-    private final String name;
-
-    public Player(String name) {
-        this.name = name;
-    }
-}
